фикс работы с mysql

+++ src/database/models/guild.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Integer
+from sqlalchemy import Column, DateTime, Integer, BigInteger
 from sqlalchemy.orm import DeclarativeBase
 
 
@@ -9,7 +9,7 @@ class Base(DeclarativeBase):
 class Guild(Base):
     __tablename__ = "guilds"
 
-    guild_id = Column(Integer, primary_key=True)
+    guild_id = Column(BigInteger, primary_key=True)
     color = Column(Integer)
-    channel_id = Column(Integer)
+    channel_id = Column(BigInteger)
     date = Column(DateTime)
